+++ CHANGELOG.md
@@ -6,12 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+-   Automatic argument validation.
 -   Arguments & options attributes for automatic validation.
 -   NoValidate attribute indicating that no automatic validation should be performed.
 -   New methods to LogHelper: Error, InvalidArgumentType.
 
 ### Changed
 
+-   Execute method now also accepts converted arguments as an additional parameter.
 -   Attributes, interfaces and enums have been moved to their respective folders and namespaces.
 -   Attributes are no longer partial, but are sealed.
 -   Templates have been updated.